Revert to less buggy version.
The PRIME Synchronization support introduced in the previous release caused certain regions under the mouse to fail to update properly under Budgie (namely Mutter). It is possible that an upstream change to reintroduce clip_region in Mutter post 3.36.3 will resolve this issue, however for now let's go ahead and revert so shannon users don't have to live with this for another week.